Divided the code into two files.
